Fix update loop for the gateway service
nginx really wants to manage all of the annotations for the created service, but metallb also wants to set the ip-allocated-from-pool annotation. This leads to a loop where metallb adds the annotation and nginx removes it again. This causes massive amounts of resource updates, which in turn causes a lot of writes and updates to external-mdns.